move flux actions to executor
