+++ src/radical/pilot/agent/scheduler/flux.py
@@ -3,46 +3,15 @@ __copyright__ = 'Copyright 2017, http://radical.rutgers.edu'
 __license__   = 'MIT'
 
 
-import os
-import shlex
-import textwrap
-
-import radical.utils        as ru
-
-from ...   import states    as rps
-from ...   import constants as rpc
-
-from ..launch_method import LaunchMethod
-from .base           import AgentSchedulingComponent
+from .noop import Noop
 
 
 # ------------------------------------------------------------------------------
 #
-class Flux(AgentSchedulingComponent):
+class Flux(Noop):
     '''
-    Pass all scheduling and execution control to Flux, and leave it to the
-    execution component to collect state updates.
+    Pass all scheduling and execution control to the Flux executor
     '''
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, session):
-
-        self.nodes = None
-
-        AgentSchedulingComponent.__init__(self, cfg, session)
-
-        # create execution helper, i.e., a small shell script which manages I/O
-        # redirection for us (Flux does not support it just yet)
-        self._helper = textwrap.dedent('''\
-                #/bin/sh
-
-                exec >>%(log)s 2>&1
-
-                %(cmd)s 1>"%(out)s" 2>"%(err)s"
-
-                ''')
-
 
     # --------------------------------------------------------------------------
     #
@@ -59,117 +28,6 @@ class Flux(AgentSchedulingComponent):
         # this abstract method is ignored in this implementation
         pass
 
-
-    # --------------------------------------------------------------------------
-    #
-    def initialize(self):
-
-        # register task output channels in *this* process
-        self.register_output(rps.AGENT_EXECUTING_PENDING,
-                             rpc.AGENT_EXECUTING_QUEUE)
-
-        super().initialize()
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        lm_cfg  = self.session.rcfg.launch_methods.get('FLUX')
-        lm_cfg['pid']       = self.session.cfg.pid
-        lm_cfg['reg_addr']  = self.session.cfg.reg_addr
-        self._lm            = LaunchMethod.create('FLUX', lm_cfg,
-                                                  self.session.cfg,
-                                                  self._log, self._prof)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def work(self, tasks):
-
-        # overload the base class work method
-        self._log.debug('submit tasks?')
-        self.advance(tasks, rps.AGENT_SCHEDULING, publish=True, push=False)
-
-        # FIXME: need actual job description, obviously
-        jds = [self.task_to_spec(task) for task in tasks]
-        self._log.debug('submit tasks: %s', [jd for jd in jds])
-        jids = self._lm.fh.submit_jobs([jd for jd in jds])
-        self._log.debug('submitted tasks')
-
-        for task, jid in zip(tasks, jids):
-
-            self._log.debug('submitted task %s -> %s', task['uid'], jid)
-
-            md = task['description'].get('metadata') or dict()
-            md['flux_id'] = jid
-            task['description']['metadata'] = md
-
-        self.advance(tasks, rps.AGENT_EXECUTING_PENDING, publish=True, push=True)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def task_to_spec(self, task):
-
-        td     = task['description']
-        uid    = task['uid']
-        sbox   = task['task_sandbox_path']
-        stdout = td.get('stdout') or '%s/%s.out' % (sbox, uid)
-        stderr = td.get('stderr') or '%s/%s.err' % (sbox, uid)
-
-        task['stdout'] = ''
-        task['stderr'] = ''
-
-        task['stdout_file'] = stdout
-        task['stderr_file'] = stderr
-
-        args = ' '.join([shlex.quote(arg) for arg in td['arguments']])
-        cmd  = '%s %s' % (td['executable'], args)
-
-        ru.rec_makedir(sbox)
-        exec_script = '%s/%s.flux.sh' % (sbox, uid)
-        with ru.ru_open(exec_script, 'w') as fout:
-            fout.write(self._helper % {'out': stdout,
-                                       'err': stderr,
-                                       'log': '%s.flux.log' % uid,
-                                       'cmd': cmd})
-        os.chmod(exec_script, 0o0755)
-        spec = {
-            'tasks': [{
-                'slot' : 'task',
-                'count': {
-                    'per_slot': 1
-                },
-                'command': [exec_script],
-            }],
-            'attributes': {
-                'system': {
-                    'cwd'     : sbox,
-                    'duration': 0,
-                }
-            },
-            'version': 1,
-            'resources': [{
-                'count': td['ranks'],
-                'type' : 'slot',
-                'label': 'task',
-                'with' : [{
-                    'count': td['cores_per_rank'],
-                    'type' : 'core'
-              # }, {
-              #     'count': td['gpus_per_rank'] or 0,
-              #     'type' : 'gpu'
-                }]
-            }]
-        }
-
-        if td['gpus_per_rank']:
-            spec['resources'][0]['with'].append({
-                    'count': td['gpus_per_rank'],
-                    'type' : 'gpu'})
-
-        return spec
 
 # ------------------------------------------------------------------------------
